IMKCandidates // Reorganization.
